rt: stabilize `Builder::rng_seed` and `RngSeed`
Moves `Builder::rng_seed`, the `runtime::RngSeed` re-export, and
`RngSeed::from_bytes` out of `cfg_unstable!`. The underlying
`seed_generator` field and runtime plumbing were already unconditional;
only the public setter, type re-export, and constructor were gated.

This lets callers seed the scheduler RNG to get deterministic `select!`
branch ordering on stable.

Closes #4879.

:house: Remote-Dev: homespace
